feat: prepare for RSC support (WIP)
Fixes #36, #38
