+++ packages/react/src/components/Button/Button.tsx
@@ -1,4 +1,3 @@
-import { ReactNode, useState } from 'react';
 import { ButtonProps as HtmlButtonProps } from 'react-html-props';
 import { twMerge } from 'tailwind-merge';
 
@@ -80,37 +79,5 @@ export const Button = ({
 		<button type="button" className={styles} {...props}>
 			{children}
 		</button>
-	);
-};
-
-export type IconButtonProps = Omit<ButtonProps, 'isIconOnly'> & {
-	'aria-label': string,
-};
-export const IconButton = ({ children, ...props }: IconButtonProps) => {
-	return (
-		<Button isIconOnly {...props}>
-			{children}
-		</Button>
-	);
-};
-
-export type ToggleButtonProps = Omit<ButtonProps, 'children'> & {
-	contentOn: ReactNode,
-	contentOff: ReactNode,
-};
-export const ToggleButton = ({ contentOn, contentOff, ...props }: ToggleButtonProps) => {
-	const [selected, setSelected] = useState(true);
-
-	return (
-		<Button
-			role='checkbox'
-			aria-checked={selected}
-			onClick={() => setSelected(selected => !selected)}
-			{...props}
-		>
-			{selected
-				? contentOn
-				: contentOff}
-		</Button>
 	);
 };
