+++ packages/react/src/components/Button/IconButton.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { IconPlaceholder } from '@tabler/icons-react';
 
-import { IconButton } from '../components/';
+import { IconButton } from '.';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta: Meta<typeof IconButton> = {
