+++ packages/react/src/components/Button/index.ts
@@ -1,0 +1,4 @@
+export type { ButtonIntent, ButtonKind, ButtonProps, ButtonSize } from './Button';
+export { Button } from './Button';
+export { IconButton, type IconButtonProps } from './IconButton';
+export { ToggleButton, type ToggleButtonProps } from './ToggleButton';
