+++ packages/react/src/components/Callout/index.ts
@@ -1,0 +1,2 @@
+export type { CalloutKind, CalloutProps } from './Callout';
+export { Callout } from './Callout';
