+++ packages/react/src/components/Checkbox/index.ts
@@ -1,0 +1,2 @@
+export type { CheckboxProps } from './Checkbox';
+export { Checkbox } from './Checkbox';
