+++ packages/react/src/components/Dropdown/Dropdown.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { IconCommand } from '@tabler/icons-react';
 
-import { Dropdown, DropdownItem, DropdownItemKbd } from '../components/';
+import { Dropdown, DropdownItem, DropdownItemKbd } from '..';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta: Meta<typeof Dropdown> = {
