+++ packages/react/src/components/Dropdown/index.ts
@@ -1,0 +1,8 @@
+export type { DropdownProps } from './Dropdown';
+export type { DropdownItemProps } from './Dropdown';
+export type { DropdownItemKbdProps } from './Dropdown';
+export type { DropdownHeadingProps } from './Dropdown';
+export { Dropdown } from './Dropdown';
+export { DropdownHeading } from './Dropdown';
+export { DropdownItem } from './Dropdown';
+export { DropdownItemKbd } from './Dropdown';
