+++ packages/react/src/components/Radio/index.ts
@@ -1,0 +1,2 @@
+export type { RadioProps } from './Radio';
+export { Radio } from './Radio';
